Add new user popup is now a separate component
